Refactor the structure of Delta Sharing provider
