+++ airflow/providers/delta/sharing/provider.yaml
@@ -35,21 +35,21 @@ integrations:
 operators:
   - integration-name: Delta Sharing
     python-modules:
-      - airflow.providers.delta_sharing.operators.delta_sharing
+      - airflow.providers.delta.sharing.operators.delta_sharing
 
 sensors:
   - integration-name: Delta Sharing
     python-modules:
-      - airflow.providers.delta_sharing.sensors.delta_sharing
+      - airflow.providers.delta.sharing.sensors.delta_sharing
 
 hooks:
   - integration-name: Delta Sharing
     python-modules:
-      - airflow.providers.delta_sharing.hooks.delta_sharing
+      - airflow.providers.delta.sharing.hooks.delta_sharing
 
 hook-class-names:  # deprecated - to be removed after providers add dependency on Airflow 2.2.0+
-  - airflow.providers.delta_sharing.hooks.delta_sharing.DeltaSharingHook
+  - airflow.providers.delta.sharing.hooks.delta_sharing.DeltaSharingHook
 
 connection-types:
-  - hook-class-name: airflow.providers.delta_sharing.hooks.delta_sharing.DeltaSharingHook
+  - hook-class-name: airflow.providers.delta.sharing.hooks.delta_sharing.DeltaSharingHook
     connection-type: delta_sharing
