+++ docs/apache-airflow/extra-packages-ref.rst
@@ -170,7 +170,7 @@ These are extras that add dependencies needed for integration with external serv
 +---------------------+-----------------------------------------------------+-----------------------------------------------------+
 | dbt.cloud           | ``pip install 'apache-airflow[dbt.cloud]'``         | dbt Cloud hooks and operators                       |
 +---------------------+-----------------------------------------------------+-----------------------------------------------------+
-| delta_sharing       | ``pip install 'apache-airflow[delta_sharing]'``     | Delta Sharing hooks, sensors and operators          |
+| delta.sharing       | ``pip install 'apache-airflow[delta.sharing]'``     | Delta Sharing hooks, sensors and operators          |
 +---------------------+-----------------------------------------------------+-----------------------------------------------------+
 | dingding            | ``pip install 'apache-airflow[dingding]'``          | Dingding hooks and sensors                          |
 +---------------------+-----------------------------------------------------+-----------------------------------------------------+
