+++ tests/providers/delta/sharing/__init__.py
